Fixed issues with writing Avro values; better tests

+++ include/rfl/avro/schema/Type.hpp
@@ -6,7 +6,7 @@
 #include <string>
 
 #include "../../Literal.hpp"
-//#include "../../Object.hpp"
+// #include "../../Object.hpp"
 #include "../../Ref.hpp"
 #include "../../Rename.hpp"
 #include "../../Variant.hpp"
@@ -36,7 +36,7 @@ struct RFL_API Type {
   };
 
   struct Double {
-    Literal<"float"> type{};
+    Literal<"double"> type{};
   };
 
   struct Bytes {
